clean up some things
